Commmit contains changes in all of the content up to Section 5.
This commit includes a substantial number of changes to the paper.
For instance, all of the new equations are now in the paper.  The
formatting of the paragraphs seems acceptable after scanning through
everything on the screen.  The next few revisions will include changes
to the material about the experimental results and analysis.  (Note
that I did make changes to remove \subsubs in Section 5).