Made modifications to the content in the first two sections of results.
This version of the paper contains changes to the content in the
empirical results section.  In particular, I made many changes to the
material about the initial response times and the vector size
variations.  There are still more marked changes to this content
and then I will modify the material about jvm behavior.